+++ CHANGELOG.md
@@ -9,13 +9,22 @@ Intended to follow [«Keep a Changelog»](https://keepachangelog.com/en/)
 
 ### TODO
 
-- Refactor `NowFactory` to PSR-20 [BREAKING] 
+- Refactor `NowFactory` to PSR-20 [BREAKING]
 - Revisit `\Brnc\CachePsr16Adapter\CacheItemPool::getTimeToLive`
-- Also `NowFactory` also for `CacheItem`
-- Add tests to ensure PSR-6 commits make use of PSR-16's bulk setters
-- Increase test coverage (using PSR-16 returning false & exceptions)
 
 ----
+
+## [1.1.0] - 2021-10-15
+
+### Changed
+
+- Change `getItems` to use `getMultiple`
+- Using `NowFactory` also for `CacheItem`
+
+### Added
+
+- Increase test coverage (using PSR-16 exceptions)
+- Add tests to ensure PSR-6 commits make use of PSR-16's bulk setters
 
 ## [1.0.0] - 2021-10-10